+++ CHANGELOG.md
@@ -21,6 +21,9 @@ The types of changes are:
 
 * Initial system management page [#1054](https://github.com/ethyca/fides/pull/1054)
 
+### Changed
+
+* Deleting a taxonomy field with children will now cascade delete all of its children as well. [#1042](https://github.com/ethyca/fides/pull/1042)
 ### Fixed
 
 * Fixed navigating directly to frontend routes loading index page instead of the correct static page for the route.